♻️ Extract step.go from step_options.go
